live-feed(js, scss): refactor, items that were for "Projects" are now "Live Feed Projects".
