+++ src/sass/components/_button.scss
@@ -1,5 +1,5 @@
-// Project Buttons
-#projects {
+// live Buttons
+#live-feed {
   button {
     border: 0;
     height: 2.5rem;
