+++ src/sass/main.scss
@@ -23,3 +23,4 @@
 @import './pages/archives';
 @import './pages/connect';
 @import './pages/skills';
+@import './pages/feed';
