+++ src/sass/pages/_feed.scss
@@ -1,0 +1,51 @@
+#live-feed {
+    height: inherit;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    color: $color-primary;
+  }
+  
+  .live-feed-page-turner {
+    margin-top: 1rem;
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+  }
+  
+  .live-feed-sort-controller {
+    display: flex;
+    width: 100%;
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: center;
+    background-color: $color-grey-dark;
+  }
+  
+  .live-feed-projects-container {
+    height: 60rem;
+    width: 100vw;
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    overflow-y: scroll;
+  
+    @include lg {
+      flex-direction: row;
+      overflow-x: scroll;
+    }
+  
+    @include xl {
+      justify-content: center;
+    }
+  
+  
+  }
+
